+++ custom_components/mydolphin_plus/common/entity_descriptions.py
@@ -41,27 +41,79 @@ from homeassistant.components.vacuum import (
     StateVacuumEntityDescription,
     VacuumEntityFeature,
 )
-from homeassistant.const import SIGNAL_STRENGTH_DECIBELS, EntityCategory, UnitOfTime
+from homeassistant.const import (
+    SIGNAL_STRENGTH_DECIBELS,
+    EntityCategory,
+    Platform,
+    UnitOfTime,
+)
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.util import slugify
 
 
 @dataclass(slots=True)
-class MyDolphinPlusVacuumEntityDescription(StateVacuumEntityDescription):
+class MyDolphinPlusEntityDescription(EntityDescription):
+    platform: Platform | None = None
+
+
+@dataclass(slots=True)
+class MyDolphinPlusVacuumEntityDescription(
+    StateVacuumEntityDescription, MyDolphinPlusEntityDescription
+):
     """A class that describes vacuum entities."""
 
+    platform: Platform | None = Platform.VACUUM
     features: VacuumEntityFeature = VacuumEntityFeature(0)
     fan_speed_list: list[str] = ()
 
 
 @dataclass(slots=True)
-class MyDolphinPlusDailyBinarySensorEntityDescription(BinarySensorEntityDescription):
+class MyDolphinPlusBinarySensorEntityDescription(
+    BinarySensorEntityDescription, MyDolphinPlusEntityDescription
+):
+    platform: Platform | None = Platform.BINARY_SENSOR
+    on_value: str | bool | None = None
+    attributes: list[str] | None = None
+
+
+@dataclass(slots=True)
+class MyDolphinPlusSensorEntityDescription(
+    SensorEntityDescription, MyDolphinPlusEntityDescription
+):
+    platform: Platform | None = Platform.SENSOR
+
+
+@dataclass(slots=True)
+class MyDolphinPlusSelectEntityDescription(
+    SelectEntityDescription, MyDolphinPlusEntityDescription
+):
+    platform: Platform | None = Platform.SELECT
+
+
+@dataclass(slots=True)
+class MyDolphinPlusNumberEntityDescription(
+    NumberEntityDescription, MyDolphinPlusEntityDescription
+):
+    platform: Platform | None = Platform.NUMBER
+
+
+@dataclass(slots=True)
+class MyDolphinPlusLightEntityDescription(
+    LightEntityDescription, MyDolphinPlusEntityDescription
+):
+    platform: Platform | None = Platform.LIGHT
+
+
+@dataclass(slots=True)
+class MyDolphinPlusDailyBinarySensorEntityDescription(
+    MyDolphinPlusBinarySensorEntityDescription
+):
     """A class that describes vacuum entities."""
 
     day: str | None = None
 
 
-ENTITY_DESCRIPTIONS: list[EntityDescription] = [
+ENTITY_DESCRIPTIONS: list[MyDolphinPlusEntityDescription] = [
     MyDolphinPlusVacuumEntityDescription(
         key=slugify(DATA_KEY_VACUUM),
         name="",
@@ -69,20 +121,20 @@ ENTITY_DESCRIPTIONS: list[EntityDescription] = [
         fan_speed_list=list(CLEANING_MODES_SHORT.keys()),
         translation_key=slugify(DATA_KEY_VACUUM),
     ),
-    LightEntityDescription(
+    MyDolphinPlusLightEntityDescription(
         key=slugify(DATA_KEY_LED),
         name=DATA_KEY_LED,
         entity_category=EntityCategory.CONFIG,
         translation_key=slugify(DATA_KEY_LED),
     ),
-    SelectEntityDescription(
+    MyDolphinPlusSelectEntityDescription(
         key=slugify(DATA_KEY_LED_MODE),
         name=DATA_KEY_LED_MODE,
         options=list(ICON_LED_MODES.keys()),
         entity_category=EntityCategory.CONFIG,
         translation_key=slugify(DATA_KEY_LED_MODE),
     ),
-    NumberEntityDescription(
+    MyDolphinPlusNumberEntityDescription(
         key=slugify(DATA_KEY_LED_INTENSITY),
         name=DATA_KEY_LED_INTENSITY,
         native_min_value=0,
@@ -91,13 +143,13 @@ ENTITY_DESCRIPTIONS: list[EntityDescription] = [
         device_class=NumberDeviceClass.POWER_FACTOR,
         translation_key=slugify(DATA_KEY_LED_INTENSITY),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_STATUS),
         name=DATA_KEY_STATUS,
         entity_category=EntityCategory.DIAGNOSTIC,
         translation_key=slugify(DATA_KEY_STATUS),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_RSSI),
         name=DATA_KEY_RSSI,
         entity_category=EntityCategory.DIAGNOSTIC,
@@ -105,48 +157,48 @@ ENTITY_DESCRIPTIONS: list[EntityDescription] = [
         native_unit_of_measurement=SIGNAL_STRENGTH_DECIBELS,
         translation_key=slugify(DATA_KEY_RSSI),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_NETWORK_NAME),
         name=DATA_KEY_NETWORK_NAME,
         entity_category=EntityCategory.DIAGNOSTIC,
         translation_key=slugify(DATA_KEY_NETWORK_NAME),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_CLEAN_MODE),
         name=DATA_KEY_CLEAN_MODE,
         entity_category=EntityCategory.DIAGNOSTIC,
         translation_key=slugify(DATA_KEY_CLEAN_MODE),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_MAIN_UNIT_STATUS),
         name=DATA_KEY_MAIN_UNIT_STATUS,
         entity_category=EntityCategory.DIAGNOSTIC,
         translation_key=slugify(DATA_KEY_MAIN_UNIT_STATUS),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_ROBOT_STATUS),
         name=DATA_KEY_ROBOT_STATUS,
         entity_category=EntityCategory.DIAGNOSTIC,
         translation_key=slugify(DATA_KEY_ROBOT_STATUS),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_ROBOT_TYPE),
         name=DATA_KEY_ROBOT_TYPE,
         entity_category=EntityCategory.DIAGNOSTIC,
         translation_key=slugify(DATA_KEY_ROBOT_TYPE),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_CYCLE_COUNT),
         name=DATA_KEY_CYCLE_COUNT,
         entity_category=EntityCategory.DIAGNOSTIC,
         translation_key=slugify(DATA_KEY_CYCLE_COUNT),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_FILTER_STATUS),
         name=DATA_KEY_FILTER_STATUS,
         translation_key=slugify(DATA_KEY_FILTER_STATUS),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_CYCLE_TIME),
         name=DATA_KEY_CYCLE_TIME,
         device_class=SensorDeviceClass.DURATION,
@@ -154,7 +206,7 @@ ENTITY_DESCRIPTIONS: list[EntityDescription] = [
         native_unit_of_measurement=UnitOfTime.MINUTES,
         translation_key=slugify(DATA_KEY_CYCLE_TIME),
     ),
-    SensorEntityDescription(
+    MyDolphinPlusSensorEntityDescription(
         key=slugify(DATA_KEY_CYCLE_TIME_LEFT),
         name=DATA_KEY_CYCLE_TIME_LEFT,
         device_class=SensorDeviceClass.DURATION,
@@ -162,7 +214,7 @@ ENTITY_DESCRIPTIONS: list[EntityDescription] = [
         native_unit_of_measurement=UnitOfTime.SECONDS,
         translation_key=slugify(DATA_KEY_CYCLE_TIME_LEFT),
     ),
-    BinarySensorEntityDescription(
+    MyDolphinPlusBinarySensorEntityDescription(
         key=slugify(DATA_KEY_AWS_BROKER),
         name=DATA_KEY_AWS_BROKER,
         icon="mdi:aws",
@@ -170,7 +222,7 @@ ENTITY_DESCRIPTIONS: list[EntityDescription] = [
         entity_category=EntityCategory.DIAGNOSTIC,
         translation_key=slugify(DATA_KEY_AWS_BROKER),
     ),
-    BinarySensorEntityDescription(
+    MyDolphinPlusBinarySensorEntityDescription(
         key=slugify(DATA_KEY_WEEKLY_SCHEDULER),
         name=DATA_KEY_WEEKLY_SCHEDULER,
         translation_key=slugify(DATA_KEY_WEEKLY_SCHEDULER),