+++ custom_components/mydolphin_plus/light.py
@@ -2,7 +2,7 @@ from abc import ABC
 import logging
 from typing import Any
 
-from homeassistant.components.light import LightEntity, LightEntityDescription
+from homeassistant.components.light import LightEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant, callback
@@ -16,6 +16,7 @@ from .common.consts import (
     ATTR_IS_ON,
     SIGNAL_DEVICE_NEW,
 )
+from .common.entity_descriptions import MyDolphinPlusLightEntityDescription
 from .managers.coordinator import MyDolphinPlusCoordinator
 
 _LOGGER = logging.getLogger(__name__)
@@ -33,7 +34,6 @@ async def async_setup_entry(
             hass,
             entry,
             Platform.LIGHT,
-            LightEntityDescription,
             MyDolphinPlusLightEntity,
             async_add_entities,
         )
@@ -48,7 +48,7 @@ class MyDolphinPlusLightEntity(MyDolphinPlusBaseEntity, LightEntity, ABC):
 
     def __init__(
         self,
-        entity_description: LightEntityDescription,
+        entity_description: MyDolphinPlusLightEntityDescription,
         coordinator: MyDolphinPlusCoordinator,
     ):
         super().__init__(entity_description, coordinator)